+++ app/root.jsx
@@ -15,7 +15,7 @@ import { ToastContainer, toast, Flip } from 'react-toastify'
 import { HomeMeta } from '~/classes/meta'
 import { NavigationBar } from '~/classes/navigationBar'
 import { Footer } from '~/classes/footer'
-import Warning from '~/components/Warning'
+import UnderConstruction from '~/components/UnderConstruction'
 
 import 'react-toastify/dist/ReactToastify.css'
 import '~/styles/styles.css'
@@ -58,8 +58,6 @@ export default function App() {
     console.log('legacy browser detected')
   }
 
-  const [warningVisible, setWarningVisible] = useState(true)
-
   useEffect(() => {
     let lastScrollTop = 0
     const mobileMenu = document.getElementById('mobile-menu')
@@ -76,10 +74,10 @@ export default function App() {
 
     // 🚧 Under Construction warning
     // Show a warning toast if it hasn't been shown in the last hour
-    const lastShown = localStorage.getItem('warningLastShown')
+    const lastShown = localStorage.getItem('UnderConstructionLastShown')
     if (!lastShown || Date.now() - parseInt(lastShown, 10) > 1000 * 60 * 60) {
-      localStorage.setItem('warningLastShown', Date.now().toString())
-      toast(Warning, {
+      localStorage.setItem('UnderConstructionLastShown', Date.now().toString())
+      toast(UnderConstruction, {
         type: 'warning',
         position: 'top-center',
         autoClose: 10000,
